Move graph location
